+++ tests/integration/reqs/test_account_tx.py
@@ -1,14 +1,10 @@
-try:
-    from unittest import IsolatedAsyncioTestCase
-except ImportError:
-    from aiounittest import AsyncTestCase as IsolatedAsyncioTestCase
-
+from tests.integration.integration_test_case import IntegrationTestCase
 from tests.integration.it_utils import test_async_and_sync
 from tests.integration.reusable_values import WALLET
 from xrpl.models.requests import AccountTx
 
 
-class TestAccountTx(IsolatedAsyncioTestCase):
+class TestAccountTx(IntegrationTestCase):
     @test_async_and_sync(globals())
     async def test_basic_functionality(self, client):
         response = await client.request(