+++ tests/integration/reqs/test_channel_verify.py
@@ -1,14 +1,10 @@
-try:
-    from unittest import IsolatedAsyncioTestCase
-except ImportError:
-    from aiounittest import AsyncTestCase as IsolatedAsyncioTestCase
-
+from tests.integration.integration_test_case import IntegrationTestCase
 from tests.integration.it_utils import test_async_and_sync
 from tests.integration.reusable_values import PAYMENT_CHANNEL, WALLET
 from xrpl.models.requests import ChannelVerify
 
 
-class TestChannelVerify(IsolatedAsyncioTestCase):
+class TestChannelVerify(IntegrationTestCase):
     @test_async_and_sync(globals())
     async def test_basic_functionality(self, client):
         response = await client.request(