+++ tests/integration/transactions/test_check_create.py
@@ -1,8 +1,4 @@
-try:
-    from unittest import IsolatedAsyncioTestCase
-except ImportError:
-    from aiounittest import AsyncTestCase as IsolatedAsyncioTestCase
-
+from tests.integration.integration_test_case import IntegrationTestCase
 from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
 from tests.integration.reusable_values import DESTINATION, WALLET
 from xrpl.models.response import ResponseStatus
@@ -15,7 +11,7 @@ EXPIRATION = 970113521
 INVOICE_ID = "6F1DFD1D0FE8A32E40E1F2C05CF1C15545BAB56B617F9C6C2D63A6B704BEF59B"
 
 
-class TestCheckCreate(IsolatedAsyncioTestCase):
+class TestCheckCreate(IntegrationTestCase):
     @test_async_and_sync(globals())
     async def test_all_fields(self, client):
         check_create = CheckCreate(