+++ tests/integration/transactions/test_deposit_preauth.py
@@ -1,8 +1,4 @@
-try:
-    from unittest import IsolatedAsyncioTestCase
-except ImportError:
-    from aiounittest import AsyncTestCase as IsolatedAsyncioTestCase
-
+from tests.integration.integration_test_case import IntegrationTestCase
 from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
 from tests.integration.reusable_values import WALLET
 from xrpl.models.response import ResponseStatus
@@ -12,7 +8,7 @@ ACCOUNT = WALLET.classic_address
 ADDRESS = "rEhxGqkqPPSxQ3P25J66ft5TwpzV14k2de"
 
 
-class TestDepositPreauth(IsolatedAsyncioTestCase):
+class TestDepositPreauth(IntegrationTestCase):
     @test_async_and_sync(globals())
     async def test_authorize(self, client):
         deposit_preauth = DepositPreauth(