+++ tests/integration/transactions/test_escrow_cancel.py
@@ -1,8 +1,4 @@
-try:
-    from unittest import IsolatedAsyncioTestCase
-except ImportError:
-    from aiounittest import AsyncTestCase as IsolatedAsyncioTestCase
-
+from tests.integration.integration_test_case import IntegrationTestCase
 from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
 from tests.integration.reusable_values import WALLET
 from xrpl.models.response import ResponseStatus
@@ -13,7 +9,7 @@ OWNER = "rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn"
 OFFER_SEQUENCE = 7
 
 
-class TestEscrowCancel(IsolatedAsyncioTestCase):
+class TestEscrowCancel(IntegrationTestCase):
     @test_async_and_sync(globals())
     async def test_all_fields(self, client):
         escrow_cancel = EscrowCancel(