+++ tests/unit/models/test_base_model.py
@@ -10,11 +10,15 @@ from xrpl.models.requests import (
     PathFind,
     PathFindSubcommand,
     PathStep,
+    Request,
     Sign,
+    SubmitMultisigned,
+    SubmitOnly,
 )
 from xrpl.models.transactions import (
     CheckCreate,
     Memo,
+    Signer,
     SignerEntry,
     SignerListSet,
     TrustSet,
@@ -59,6 +63,22 @@ class TestBaseModel(TestCase):
         )
         self.assertEqual(repr(amount), expected_repr)
 
+    def test_is_dict_of_model_when_true(self):
+        self.assertTrue(
+            IssuedCurrencyAmount.is_dict_of_model(
+                IssuedCurrencyAmount.from_dict(amount_dict).to_dict(),
+            ),
+        )
+
+    def test_is_dict_of_model_when_not_true(self):
+        self.assertFalse(
+            Sign.is_dict_of_model(
+                IssuedCurrencyAmount.from_dict(amount_dict).to_dict(),
+            ),
+        )
+
+
+class TestFromDict(TestCase):
     def test_from_dict_basic(self):
         amount = IssuedCurrencyAmount.from_dict(amount_dict)
         self.assertEqual(amount, IssuedCurrencyAmount(**amount_dict))
@@ -276,6 +296,91 @@ class TestBaseModel(TestCase):
         with self.assertRaises(XRPLModelException):
             SignerListSet.from_dict(dictionary)
 
+    def test_from_dict_multisign(self):
+        txn_sig1 = (
+            "F80E201FE295AA08678F8542D8FC18EA18D582A0BD19BE77B9A24479418ADBCF4CAD28E7BD"
+            "96137F88DE7736827C7AC6204FBA8DDADB7394E6D704CD1F4CD609"
+        )
+        txn_sig2 = (
+            "036E95B8100EBA2A4A447A3AF24500261BF480A0E8D62EE15D03A697C85E73237A5202BD9A"
+            "F2D9C68B8E8A5FA8B8DA4F8DABABE95E8401C5E57EC783291EF80C"
+        )
+        pubkey1 = "ED621D6D4FF54E809397195C4E24EF05E8500A7CE45CDD211F523A892CDBCDCDB2"
+        pubkey2 = "EDD3ABCFF008ECE9ED3073B41913619341519BFF01F07331B56E5D6D2EC4A94A57"
+        request = {
+            "method": "submit_multisigned",
+            "tx_json": {
+                "Account": "rnD6t3JF9RTG4VgNLoc4i44bsQLgJUSi6h",
+                "TransactionType": "TrustSet",
+                "Fee": "10",
+                "Sequence": 17896798,
+                "Flags": 131072,
+                "Signers": [
+                    {
+                        "Signer": {
+                            "Account": "rGoKUCwJ2C4oHUsqnd8PVxZhFiBMV2T42G",
+                            "TxnSignature": txn_sig1,
+                            "SigningPubKey": pubkey1,
+                        }
+                    },
+                    {
+                        "Signer": {
+                            "Account": "rsi3GL27pstEYUJ28ZM3q155rmFCCTBCZ1",
+                            "TxnSignature": txn_sig2,
+                            "SigningPubKey": pubkey2,
+                        }
+                    },
+                ],
+                "SigningPubKey": "",
+                "LimitAmount": {
+                    "currency": "USD",
+                    "issuer": "rH5gvkKxGHrFAMAACeu9CB3FMu7pQ9jfZm",
+                    "value": "10",
+                },
+            },
+            "fail_hard": False,
+        }
+        expected = SubmitMultisigned(
+            tx_json=TrustSet(
+                account="rnD6t3JF9RTG4VgNLoc4i44bsQLgJUSi6h",
+                fee="10",
+                sequence=17896798,
+                flags=131072,
+                signers=[
+                    Signer(
+                        account="rGoKUCwJ2C4oHUsqnd8PVxZhFiBMV2T42G",
+                        txn_signature=txn_sig1,
+                        signing_pub_key=pubkey1,
+                    ),
+                    Signer(
+                        account="rsi3GL27pstEYUJ28ZM3q155rmFCCTBCZ1",
+                        txn_signature=txn_sig2,
+                        signing_pub_key=pubkey2,
+                    ),
+                ],
+                limit_amount=IssuedCurrencyAmount(
+                    currency="USD",
+                    issuer="rH5gvkKxGHrFAMAACeu9CB3FMu7pQ9jfZm",
+                    value="10",
+                ),
+            ),
+        )
+        actual = Request.from_dict(request)
+        self.assertEqual(actual, expected)
+
+    def test_from_dict_submit(self):
+        blob = "SOISUSF9SD0839W8U98J98SF"
+        id_val = "submit_786514"
+        request = {
+            "method": "submit",
+            "tx_blob": blob,
+            "fail_hard": False,
+            "id": id_val,
+        }
+        expected = SubmitOnly(tx_blob=blob, id=id_val)
+        actual = Request.from_dict(request)
+        self.assertEqual(actual, expected)
+
     def test_from_xrpl(self):
         dirname = os.path.dirname(__file__)
         full_filename = "x-codec-fixtures.json"
@@ -293,6 +398,69 @@ class TestBaseModel(TestCase):
                     tx = Transaction.from_xrpl(r_json)
                     translated_tx = transaction_json_to_binary_codec_form(tx.to_dict())
                     self.assertEqual(r_json, translated_tx)
+
+    def test_from_xrpl_signers(self):
+        txn_sig1 = (
+            "F80E201FE295AA08678F8542D8FC18EA18D582A0BD19BE77B9A24479418ADBCF4CAD28E7BD"
+            "96137F88DE7736827C7AC6204FBA8DDADB7394E6D704CD1F4CD609"
+        )
+        txn_sig2 = (
+            "036E95B8100EBA2A4A447A3AF24500261BF480A0E8D62EE15D03A697C85E73237A5202BD9A"
+            "F2D9C68B8E8A5FA8B8DA4F8DABABE95E8401C5E57EC783291EF80C"
+        )
+        pubkey1 = "ED621D6D4FF54E809397195C4E24EF05E8500A7CE45CDD211F523A892CDBCDCDB2"
+        pubkey2 = "EDD3ABCFF008ECE9ED3073B41913619341519BFF01F07331B56E5D6D2EC4A94A57"
+        tx = {
+            "Account": "rnoGkgSpt6AX1nQxZ2qVGx7Fgw6JEcoQas",
+            "TransactionType": "TrustSet",
+            "Fee": "10",
+            "Sequence": 17892983,
+            "Flags": 131072,
+            "Signers": [
+                {
+                    "Signer": {
+                        "Account": "rGVXgBz4NraZcwi5vqpmwPW6P4y74A4YvX",
+                        "TxnSignature": txn_sig1,
+                        "SigningPubKey": pubkey1,
+                    }
+                },
+                {
+                    "Signer": {
+                        "Account": "rB5q2wsHeXdQeh2KFzBb1CujNAfSKys6GN",
+                        "TxnSignature": txn_sig2,
+                        "SigningPubKey": pubkey2,
+                    }
+                },
+            ],
+            "SigningPubKey": "",
+            "LimitAmount": {
+                "currency": "USD",
+                "issuer": "rBPvTKisx7UCGLDtiUZ6mDssXNREuVuL8Y",
+                "value": "10",
+            },
+        }
+        expected = TrustSet(
+            account="rnoGkgSpt6AX1nQxZ2qVGx7Fgw6JEcoQas",
+            fee="10",
+            sequence=17892983,
+            flags=131072,
+            signers=[
+                Signer(
+                    account="rGVXgBz4NraZcwi5vqpmwPW6P4y74A4YvX",
+                    txn_signature=txn_sig1,
+                    signing_pub_key=pubkey1,
+                ),
+                Signer(
+                    account="rB5q2wsHeXdQeh2KFzBb1CujNAfSKys6GN",
+                    txn_signature=txn_sig2,
+                    signing_pub_key=pubkey2,
+                ),
+            ],
+            limit_amount=IssuedCurrencyAmount(
+                currency="USD", issuer="rBPvTKisx7UCGLDtiUZ6mDssXNREuVuL8Y", value="10"
+            ),
+        )
+        self.assertEqual(Transaction.from_xrpl(tx), expected)
 
     def test_from_xrpl_memos(self):
         memo_type = "687474703a2f2f6578616d706c652e636f6d2f6d656d6f2f67656e65726963"
@@ -333,17 +501,3 @@ class TestBaseModel(TestCase):
             ),
         )
         self.assertEqual(Transaction.from_xrpl(tx), expected)
-
-    def test_is_dict_of_model_when_true(self):
-        self.assertTrue(
-            IssuedCurrencyAmount.is_dict_of_model(
-                IssuedCurrencyAmount.from_dict(amount_dict).to_dict(),
-            ),
-        )
-
-    def test_is_dict_of_model_when_not_true(self):
-        self.assertFalse(
-            Sign.is_dict_of_model(
-                IssuedCurrencyAmount.from_dict(amount_dict).to_dict(),
-            ),
-        )