+++ CHANGELOG.md
@@ -17,8 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Additional X-Address support in the binary codec and account sugar
 - Method to the `Wallet` class that generates the X-Address for the wallet's account
 - Async versions of all high-level functionality within the `xrpl.asyncio` package
-- Adds a web socket client under `xrpl.clients.WebsocketClient`
-- Adds an async web socket client under `xrpl.asyncio.clients.AsyncWebsocketClient`
+- Web socket client under `xrpl.clients.WebsocketClient`
+- Async web socket client under `xrpl.asyncio.clients.AsyncWebsocketClient`
+- A general `from_dict` method for `Request` models, analogous to `Transaction.from_dict`
 
 ### Fixed
 - Typos in docs/tests